handle remarks/questions from review.
