+++ invenio_vocabularies/datastreams/readers.py
@@ -20,12 +20,13 @@ from json.decoder import JSONDecodeError
 
 import requests
 import yaml
+from lxml import etree
+from lxml.html import parse as html_parse
 from oaipmh_scythe import Scythe
 from oaipmh_scythe.exceptions import NoRecordsMatch
-from lxml.html import parse as html_parse
+from oaipmh_scythe.models import Record
 
 from .errors import ReaderError
-from .models import OAIRecord
 from .xml import etree_to_dict
 
 
@@ -252,21 +253,17 @@ class OAIPMHReader(BaseReader):
 
     def _iter(self, scythe, *args, **kwargs):
         """Read and parse an OAIPMH stream to dict."""
-        scythe.class_mapping["ListRecords"] = OAIRecord
+        scythe.class_mapping["ListRecords"] = self.OAIRecord
         try:
-            responses = scythe.list_records(
+            records = scythe.list_records(
                 from_=self._from,
                 until=self._until,
                 metadata_prefix=self._metadata_prefix,
                 set_=self._set,
                 ignore_deleted=True,
             )
-            for response in responses:
-                oaipmh_record = {
-                    "header": response.header,
-                    "metadata": response.metadata["record"],
-                }
-                yield oaipmh_record
+            for record in records:
+                yield {"record": record}
         except NoRecordsMatch:
             raise ReaderError(f"No records found in OAI-PMH request.")
 
@@ -278,3 +275,31 @@ class OAIPMHReader(BaseReader):
         else:
             with Scythe(self._base_url) as scythe:
                 yield from self._iter(scythe=scythe, *args, **kwargs)
+
+    class OAIRecord(Record):
+        """An XML unpacking implementation for more complicated formats."""
+
+        def get_metadata(self):
+            """Extract and return the record's metadata as a dictionary."""
+            return xml_to_dict(
+                self.xml.find(".//" + self._oai_namespace + "metadata").getchildren()[
+                    0
+                ],
+            )
+
+
+def xml_to_dict(tree: etree._Element):
+    """Convert an XML tree to a dictionary.
+
+    This function takes an XML element tree and converts it into a dictionary.
+
+    Args:
+        tree: The root element of the XML tree to be converted.
+
+    Returns:
+        A dictionary with the key "record".
+    """
+    dict_obj = dict()
+    dict_obj["record"] = etree.tostring(tree)
+
+    return dict_obj
